fix plasp instance input

+++ consume_nogoods.py
@@ -58,21 +58,20 @@ def setup_logging(no_stream_output=False, logtofile=None):
         consoleHandler.setFormatter(formatter)
         rootLogger.addHandler(consoleHandler)
 
-
-
-def plasp_translate(instance, domain, filename):
+def plasp_translate(instance, domain, filename, no_fd):
 
     if domain is None:
         # look for domain in the same folder
         logging.info("testing domain path: {}".format(os.path.join(get_parent_dir(instance), "domain.pddl")))
         if os.path.isfile(os.path.join(get_parent_dir(instance), "domain.pddl")):
-            omain = os.path.join(get_parent_dir(instance), "domain.pddl")
-
+            domain = os.path.join(get_parent_dir(instance), "domain.pddl")
+            logging.info("Succes finding domain!")
         # look for domain in parent folder
         else:
             logging.info("testing domain path: {}".format(os.path.join(get_parent_dir(get_parent_dir(instance)), "domain.pddl"))) 
             if os.path.isfile(os.path.join(get_parent_dir(get_parent_dir(instance)), "../domain.pddl")):
                 domain = os.path.join(get_parent_dir(get_parent_dir(instance)), "../domain.pddl")
+                logging.info("Succes finding domain!")
 
             else:
                 logging.error("no domain could be found. Exiting...")
@@ -81,11 +80,18 @@ def plasp_translate(instance, domain, filename):
 
     logging.info("translating instance {}\nwith domain {}".format(instance, domain))
 
-    fd_call = config_file.FD_CALL + [domain, instance]
+    if not no_fd:
+        logging.info("Calling Fast Downward preprocessing...")
+        fd_call = config_file.FD_CALL + [domain, instance]
+        instance_files = ["output.sas"]
 
-    output = subprocess.check_output(fd_call).decode("utf-8")
+        output = subprocess.check_output(fd_call).decode("utf-8")
 
-    plasp_call = [config_file.PLASP, "translate", "output.sas"]
+    else:
+        instance_files = [domain, instance]
+
+    logging.info("Translating with plasp...")
+    plasp_call = [config_file.PLASP, "translate"] + instance_files
 
     output = subprocess.check_output(plasp_call).decode("utf-8")
     with open(filename, "w") as f:
@@ -93,7 +99,8 @@ def plasp_translate(instance, domain, filename):
 
     logging.info("saved translation into {}".format(filename))
 
-    os.remove("output.sas")
+    if not no_fd:
+        os.remove("output.sas")
 
 
 def call_clingo(file_names, time_limit, options):
@@ -274,6 +281,7 @@ if __name__ == "__main__":
     parser.add_argument("--pddl-instance", help="pddl instance")
     parser.add_argument("--pddl-domain", help="pddl domain")
     parser.add_argument("--trans-name", help="name of the translated file")
+    parser.add_argument("--no-fd", action="store_true", help="When translating the pddl instance, do not use Fast Downward preprocessing.")
 
     parser.add_argument("--save-folder", help="name of the folder where results will be written to", default=None)
     parser.add_argument("--print-results", action="store_true", help="Print results to stdout")
@@ -294,7 +302,7 @@ if __name__ == "__main__":
         else:
             trans_name = "instance-temp.lp"
 
-        plasp_translate(args.pddl_instance, args.pddl_domain, trans_name)
+        plasp_translate(args.pddl_instance, args.pddl_domain, trans_name, args.no_fd)
 
         files.append(trans_name)
